+++ src/lib/utils/fiberize.js
@@ -1,0 +1,32 @@
+import Fiber from 'fibers';
+
+// Wrap a function in a fiber.
+// Correctly handles expected presence of done callback
+export function fiberize(fn) {
+  return function (done) {
+    const self = this;
+    Fiber(function () {
+      try {
+        if (fn.length == 1) {
+          fn.call(self, done);
+        } else {
+          fn.call(self);
+          done();
+        }
+      } catch (e) {
+        process.nextTick(function () {
+          throw(e);
+        });
+      }
+    }).run();
+  };
+}
+
+export function fiberizeSync(fn) {
+  return function () {
+    const self = this;
+    Fiber(function () {
+      fn.call(self);
+    }).run();
+  };
+}